% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPossibleCombinations.R
\name{allPossibleCombinations}
\alias{allPossibleCombinations}
\title{Generate All Possible Combinations of m Out of n}
\usage{
allPossibleCombinations(n, m)
}
\arguments{
\item{n}{n}

\item{m}{m}
}
\value{
a JSON string with all possible combinations;
  object of class \code{json} as returned by \code{\link[jsonlite]{toJSON}}
}
\description{
Generate All Possible Combinations of m Out of n
}
\examples{
allPossibleCombinations(4, 2)
}
